pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "omni-tool-runtime"
version = "0.1.0"
description = "Portable tool runtime for OmniBioAI TES (RESULT_URI uploader for s3:// and azureblob://)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Manish Kumar" }]
keywords = ["omnibioai", "tes", "bioinformatics", "batch", "azure", "aws"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = []


[project.optional-dependencies]
aws   = ["boto3>=1.26"]
azure = ["azure-identity>=1.15", "azure-storage-blob>=12.19"]
dev   = [
  # test runner + coverage
  "pytest>=8.0",
  "pytest-cov>=5.0",
  # linting / formatting
  "ruff>=0.4",
  # type checking
  "mypy>=1.10",
  "boto3-stubs[s3]>=1.26",
  # azure stubs (optional but silences mypy on azure paths)
  "azure-storage-blob>=12.19",
  "azure-identity>=1.15",
]


[project.scripts]
omni-tool-runtime = "omni_tool_runtime.run:main"


[tool.setuptools.packages.find]
where    = ["."]
include  = ["omni_tool_runtime*", "tools*"]


# ── pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths    = ["tests"]
addopts      = "-v --tb=short"


# ── coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source   = ["omni_tool_runtime"]
branch   = true

[tool.coverage.report]
show_missing   = true
skip_covered   = false
fail_under     = 85
exclude_lines  = [
  "pragma: no cover",
  "if __name__ == .__main__.:",
  "raise NotImplementedError",
]


# ── ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length    = 100
target-version = "py310"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # isort
  "UP",  # pyupgrade
  "B",   # flake8-bugbear
  "SIM", # flake8-simplify
]
ignore = [
  "E501",  # line too long — handled by formatter
  "B904",  # raise-without-from-inside-except — intentional in user-facing error helpers
]

[tool.ruff.lint.per-file-ignores]
"tests/test_azureblob_uploader.py" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["omni_tool_runtime"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"


# ── mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version         = "3.12"
strict                 = false
ignore_missing_imports = true
warn_unused_ignores    = true
warn_return_any        = false
files                  = ["omni_tool_runtime"]